pkg/provider/aci_utils.go
/*
Copyright (c) Microsoft Corporation.
Licensed under the Apache 2.0 license.
*/
package provider

import (
	"crypto/sha256"
	"encoding/hex"
	"fmt"
	"strings"
)

func getContainerID(cgID, containerName *string) string {
	if cgID == nil {
		return ""
	}

	containerResourceID := fmt.Sprintf("%s/containers/%s", *cgID, *containerName)

	h := sha256.New()
	if _, err := h.Write([]byte(strings.ToUpper(containerResourceID))); err != nil {
		panic(err)
	}
	hashBytes := h.Sum(nil)
	return fmt.Sprintf("aci://%s", hex.EncodeToString(hashBytes))
}

func omitDuplicates(strs []string) []string {
	uniqueStrs := make(map[string]bool)

	var ret []string
	for _, str := range strs {
		if !uniqueStrs[str] {
			ret = append(ret, str)
			uniqueStrs[str] = true
		}
	}
	return ret
}
